Separated the validators only relevant to offender case into another concerns
