+++ app/form_models/offender_sar_complaint_case_form.rb
@@ -1,6 +1,8 @@
 module OffenderSARComplaintCaseForm
   extend ActiveSupport::Concern
-  
+
+  include OffenderFormValidators
+
   STEPS = %w[link-offender-sar-case
              confirm-offender-sar
              requester-details
